Legion.Identity.Auth.Abstract.Token (legion vdev)

Defines credential-based authentication functions in order to authenticate users.

Types

t()

Specs

t() :: %Legion.Identity.Auth.Abstract.Token{
  expires_after: pos_integer(),
  header_value: binary(),
  jwk: map(),
  jws: map()
}

Data data structure representing issue result.

Functions

issue_token(user_id, passphrase_id, opts \\ [])

Specs

issue_token(
  Legion.Identity.Information.Registration.id(),
  Legion.Identity.Auth.Concrete.Passphrase.id(),
  Keyword.t()
) :: {:ok, t()} | {:error, :invalid} | {:error, :timed_out}

Issues a token to given user.

Options

:lifetime: If supplied, the lifetime of the token will be changed to this value. This defaults to the configuration value.

Applied configuration values

On Legion.Identity.Auth.Concrete.JOSE domain,

:issuer: The issuer of the token, e.g. "legion".

:sub: The subsidiary of the token.

If :lifetime is not supplied as an option, the value of the parameter will be applicable to:

:lifetime: The default lifetime of the token.