Legion.Identity.Auth.Algorithm.Keccak (legion vdev)

Provides functions for hashing data with FIPS-202 compliant Keccak variants.

Types

algorithm()

Specs

algorithm() :: :sha3_224 | :sha3_256 | :sha3_384 | :sha3_512

Keccak variants to be used in default hashing operations.

hash()

Specs

hash() :: binary()

The result of the algorithm.

Functions

hash(data)

Specs

hash(binary()) :: hash()

Hashes the given binary with default Keccak algorithm configured.

hash(algorithm, data)

Specs

hash(algorithm(), binary()) :: hash()

Hashes the given binary with given Keccak algorithm.