Legion.Identity.Auth.Concrete (legion vdev)

Functions for performing concrete authentication.

Functions

generate_passphrase(username, password_hash, ip_addr, opts \\ [])

Specs

generate_passphrase(
  Legion.Identity.Information.Registration.username(),
  Legion.Identity.Auth.Algorithm.Keccak.hash(),
  Legion.Networking.INET.t(),
  Keyword.t()
) ::
  {:ok, :access, Legion.Identity.Auth.Abstract.Token.t()}
  | {:ok, :require, Legion.Identity.Auth.Concrete.Passkey.t()}
  | {:ok, :advance, Advance.t()}
  | {:error, :no_user_verify}
  | {:error, :unsupported_scheme}
  | {:error, :wrong_password}
  | {:error, :maximum_passphrases_exceeded}
  | {:error, Legion.Networking.INET.error_type()}

Performs concrete authentication for the user with given username, and client-side produced hash. Note that password hash is not password digest, where digest is a time-variant, salted hash of the client-side produced hash.

Options

:one_shot: If true, the authentication will be fallthrough, it will not generate any passphrases, but return a token without requiring any further abstract authentication. This is useful when user does not need to store the passphrase in the authenticating device. The lifetime of the token is determined seperately from the regular abstract authentication flow. Defaults to false.

Return types

With a valid digest for the user, this function should either return a

{:ok, :require, passkey} tuple containing the passkey which should be transferred to the client-side device to be persisted on its secure storage, or

{:ok, :access, jwt} tuple containing the JSON Web Token (RFC 7159) to perform stealth authentication directly, or

{:ok, :advance, advance_artifact} tuple containing the artifact information for proceeding to the next mandatory step of users preferred multi-factor-authentication method.

In such circumstances, these errors can also be returned from the function:

{:error, :no_user_verify}: User cannot be found with given username.

{:error, :unsupported_scheme}: The authentication scheme of the user is not supported (in implementation). This is subject to change without a notice on further releases and left here for the regression of provisioning services.

{:error, :wrong_password_hash}: The password hash is wrong.

{:error, :maximum_passphrases_exceeded}: The number of the passphrases (5) are exceeded. User might either reset his/her password, authentication will not proceed.

{:error, :bad_host, :reserve_reason}: The IP address provided is reserved by IETF (IANA).

{:error, :blacklist, "reason"}: The IP address provided is blacklisted by server authority.

Applied configuration values

If :one_shot option is enabled, the function takes Legion.Identity.Auth.Concrete.JOSE.extended_lifetime configuration value to provide an expiration time to the token issuer.

register_internal_user(username, password_hash)

Specs

register_internal_user(String.t(), String.t()) ::
  {:ok, Legion.Identity.Information.Registration.id(),
   Legion.Identity.Information.Registration.username()}
  | {:error, :already_registered}

Registers an internal user with given username and password hash.

Caveats

The new user will have the insecure authentication scheme set initially, client-side implementations are expected to ask whether user prefers another authentication scheme or not at the time of the completion of this action.