Legion.Identity.Auth.Concrete.Activity (legion vdev)

Activities represent token generation from passphrases.

Functions

changeset(struct, params \\ %{})

Specs

changeset(Legion.Identity.Auth.Concrete.Activity, map()) :: Ecto.Changeset.t()

Creates a changeset with given parameters.

create_changeset(passphrase_id, user_agent_string, ip_addr, gps_location)

Specs

create_changeset(
  Legion.Identity.Auth.Concrete.Passphrase.id(),
  Legion.Networking.HTTP.UserAgent.t(),
  Legion.Networking.INET.t(),
  Postgrex.Point.t()
) :: {:ok, Ecto.Changeset.t()} | {:error, any()}

Creates a changeset with given passphrase, user agent string and IP address.

generate_activity(passphrase, user_agent, ip_addr, gps_location)

Specs

generate_activity(
  Legion.Identity.Auth.Concrete.Passphrase,
  binary(),
  Legion.Networking.INET.t(),
  Postgrex.Point.t()
) ::
  {:ok, Legion.Identity.Auth.Concrete.Activity} | {:error, Ecto.Changeset.t()}

last_activity(user)

Specs

last_activity(Legion.Identity.Information.Registration | integer()) ::
  Legion.Identity.Auth.Concrete.Activity | nil

Fetches last activity of the user from the database. One can supply either User struct or user identifier as an integer as a parameter.