Legion.Identity.Auth.Concrete.Passphrase (legion vdev)

Passphrase (a.k.a. access token) is an artifact of a successful concrete authentication.

Types

id()

Specs

id() :: integer()

The type of the identifier to uniquely reference a passphrase.

Functions

changeset(struct, params \\ %{})

Specs

changeset(Legion.Identity.Auth.Concrete.Passphrase, map()) :: Ecto.Changeset.t()

check_passphrase_quota(user_id)

Checks passphrase quota of the user, returns error if it is exceeded.

create(user_or_id, ip_addr)

Specs

create(
  Legion.Identity.Information.Registration.user_or_id(),
  Legion.Networking.INET.t()
) :: Legion.Identity.Auth.Concrete.Passkey.t()

Creates a changeset and inserts it into the repository.

create_changeset(user_id, ip_addr)

Specs

create_changeset(
  Legion.Identity.Information.Registration.user_or_id(),
  Legion.Networking.INET.t()
) :: {Legion.Identity.Auth.Concrete.Passkey.t(), Ecto.Changeset.t()}

Generates a new passphrase changeset and returns it with the passkey itself in a tuple.

exists?(passphrase_id)

Specs

exists?(id()) :: boolean()

Returns a boolean value indicating if passphrase exists.

find_passphrase_matching(user_id, passkey)

Specs

find_passphrase_matching(
  Legion.Identity.Information.Registration.id(),
  Legion.Identity.Auth.Concrete.Passkey.t()
) ::
  {:ok, id()} | {:error, :not_found} | {:error, :invalid} | {:error, :timed_out}

Searches for a matching passphrase with given user identifier and a passkey hash.

Return types

This function returns a passphrase identifier if there was no faulty conditions. However, it might also return errors in following scenarios.

:not_found: No matching passphrase found.

:invalid: The passphrase exists, but it is revoked manually.

:timed_out: The passphrase exists, however it is not valid anymore.

The interface of the business logic might choose to not show the reason for the failed action due to security concerns. Hence, the developer of this API might need to take care of the error values implicitly.

Discussion

We are searching for the passphrases, but not using the Elixir.Legion.Identity.Auth.Concrete.ActivePassphrase view schema. The reason on doing this is, we need to offload the query conditionals to the web application instead of the server.

The passkey itself is a unique index, so it is not overwhelming to query one with this attribute on passphrases. The only thing we need to do additionally is looking for an invalidation entry, since we are using unique index here, again, we can leverage the power of the physical btree indexing the invalidations.

validate(passphrase)

Specs

validate(Legion.Identity.Auth.Concrete.Passphrase) ::
  :ok | {:error, :invalid} | {:error, :timed_out}

Validates given passphrase and returns :ok, or an error tuple with a reason.

A passphrase is valid if and only if it is not invalidated manually and it has not timed out. This function checks its invalidation is nil and its insertion time is in a viable lifetime interval. It returns {:error, :invalid} if passphrase is invalidated manually, or {:error, :timed_out} if passphrase is timed out.

Caveats

To use this function properly, one should preload :invalidation association of the passphrase before supplying it as a parameter to this function. Otherwise, nil-check is performed against the :invalidation field of the struct, it would see the association as non-nil since Ecto.Association.NotLoaded struct will be there.

validate_id(passphrase_id)

Specs

validate_id(id()) ::
  :ok | {:error, :not_found} | {:error, :invalid} | {:error, :timed_out}

Validates a passphrase with given identifier.