Legion.Identity.Auth.TFA.OneTimeCode (legion vdev)

One time codes allow users to enable two-factor authentication (2FA). One may decide to use OTCs with SMS or e-mail. OTC is server-generated, and it cannot be resembled by any client.

Types

algorithm()

Specs

algorithm() :: :argon2 | :bcrypt | :pbkdf2

format()

Specs

format() :: :integer | :alphanumeric

Type of a one-time-code.

t()

Specs

t() :: binary()

Represents resulting artifact of generator functions.

Functions

generate()

Specs

generate() :: t()

Generates a one-time-code with type declared in configuration. Use generate/1 to explicitly set the type instead.

generate(atom)

Specs

generate(format()) :: t()

Generates a one-time-code in given type.

hash(otc)

Specs

hash(t()) :: binary()

Hashes given one-time-code with the default Keccak variant declared in configuration file.

stall()

Specs

stall() :: false

A dummy stall function in order to prevent from handle enumeration. It always return false.