Legion.Identity.Information.AddressBook.Address (legion vdev)

Represents an entry found in address book.

Schema fields

:user_id: The identifier of the user that address belongs to.

:type: Type of the address, e.g. "home", "work".

:name: Name of the address, e.g. "Home at Rotterdam".

:description: Description of the address. Most of the time this is a short explanation of the address in a particular language to help to navigate the readers of the address properly.

:state: Name of the state, e.g. "Kansas". Some countries might not respect the political divisions to this type of attribute.

:city: Name of the city, e.g. "Rotterdam".

:neighborhood: Name of the neighborhood, e.g. "Beukelsdijk".

:zip_code: Zip code of the location.

Types

address_or_id()

Specs

address_or_id() :: Address.id() | Address

Represents a polymorphic type, containing either the struct itself or a unique identifier.

address_type()

Specs

address_type() :: :home | :work | :other

Shows the type of the address.

id()

Specs

id() :: pos_integer()

Unique identifier for the address.

Functions

changeset(struct, params \\ %{})