Legion.Identity.Information.PersonalData (legion vdev)

Represents personal information of the user.

Schema fields

:given_name: A name given to a individual to differentiate them from members of such group or family. May be referred as "family name" or "forename".

:middle_name: A name given to a individual to differentiate them from members of such group or family with members having same given name.

:family_name: A name given to a individual to represent its family or group.

:name_prefix: Prefix of the name to vocal the individual, such as "Mr.".

:name_postfix: Postfix of the name to vocal the individual, such as "-san", which may expand to something like "Suguro-san".

:nickname: A nick given to an individual, generally pronounced informally.

:phonetic_representation: Phonetic representation of the name of the user.

:nationality_abbreviation: Nationality of a user, e.g. "sa" for Saudi Arabian.

:gender: Gender of a user.

Functions

changeset(struct, params \\ %{})

validate_range(changeset, field, range)