Legion.Identity.Information.User (legion vdev)

Provides repository functionality for the users.

Functions

get_user(id)

Specs

get_user(integer()) ::
  {:error, :einval | :not_found}
  | {:ok, Legion.Identity.Information.Registration.t()}

Retrieves the user with the given identifier.

Parameters

id: Designating identifier value for the user.

Return values

{:ok, user}: Query succeeded, contains the user.

{:error, :not_found}: Query succeeded but user not found.

{:error, :einval}: Query failed, invalid parameter.

list_users(limit \\ 10, offset \\ 0)

Specs

list_users(integer(), integer()) ::
  {:ok, [Legion.Identity.Information.Registration.t()]} | {:error, :einval}

Lists users with the provided pagination parameters.

Parameters

limit: Limits the number of the users by the given number. Default is 10, maximum allowed value is 20.

offset: Skips the given number of preceding entries in the query. Default is 0.

Return values

{:ok, users}: Query succeeded, contains list of users.

{:error, :einval}: Query failed, invalid parameter.