Legion.Identity.Telephony.PhoneNumber (legion vdev)

Represents a phone number entry of a user.

Schema fields

:user_id: The reference of the user that phone number belongs to.

:number: The number of the phone.

:type: The type of the phone, e.g. "home", "work".

:ignored?: If true, the phone number will be hidden from transactional endpoints.

:safe?: If false, the phone number will not be used in authentication processes.

:primed_at: When the user marks the phone number entry as "primary", this attribute will be updated with a timestamp.

:inserted_at: The timestamp user created the entry.

Prioritization of entries

Phone number entries can be prioritized by a user, possibly for telephony-related transactional operations. A user may make phone number on a arbitrary moment. If no phone number of a user was made primary, the phone number will be picked according to configuration options. However, no matter what, the rules below for determining the phone number will still apply.

If a phone number is primary,

the phone number is not ignored,

it should have no descendants with a higher insertion timestamp,

it should have no descendants with a higher prioritization timestamp,

its type is one of "home", "work", "mobile".

Rules for prioritization

Certain rules apply for the prioritization of the phone numbers. A phone number entry can be prioritized if and only if,

it is not marked as ignored,

it is marked as safe.

Types

id()

Specs

id() :: pos_integer()

A positive integer uniquely identifying a phone number entity.

phone_type()

Specs

phone_type() :: :home | :mobile | :work | :home_fax | :work_fax | :pager

Type of the phone number.

Functions

changeset(struct, params \\ %{})