Legion.Identity.Telephony.PhoneNumber.FiniteStateMachine (legion vdev)

Manages phone numbers of users, performs redirections of phone calls.

This module is NOT transaction-safe.

Functions

acknowledge_phone_number(phone_number_id)

Specs

acknowledge_phone_number(Legion.Identity.Telephony.PhoneNumber.id()) ::
  {:ok, Legion.Identity.Telephony.PhoneNumber.NeglectionTrait}
  | {:ok, :noop}
  | {:error, :not_found}

Acknowledges phone number. Returns the removed neglection trait.

Same as ignore_phone_number/1, but works opposite.

create_phone_number(user_id, type, number)

Specs

create_phone_number(
  Legion.Identity.Information.Registration.id(),
  Legion.Identity.Telephony.PhoneNumber.phone_type(),
  String.t()
) ::
  {:ok, Legion.Identity.Telephony.PhoneNumber}
  | {:error, :no_user}
  | {:error, :invalid}
  | {:error, :unknown_type}

Adds a phone number entry to the user.

To get more information about the fields, see Legion.Identity.Telephony.PhoneNumber.

exists?(phone_number_id)

Specs

exists?(Legion.Identity.Telephony.PhoneNumber.id()) :: boolean()

Returns a boolean value indicating whether the phone number exists.

ignore_phone_number(passphrase_id, phone_number_id)

Specs

ignore_phone_number(
  Legion.Identity.Auth.Concrete.Passphrase.id(),
  Legion.Identity.Telephony.PhoneNumber.id()
) ::
  {:ok, Legion.Identity.Telephony.PhoneNumber.NeglectionTrait}
  | {:ok, :noop}
  | {:error, :primary}
  | {:error, :no_passphrase}
  | {:error, :not_found}

Ignores given phone number.

Return values

On successful operation, this will return {:ok, neglection_trait}.

Otherwise,

{:ok, :noop}, if the phone number was already ignored,

{:error, :not_found}, if there was no phone number with given identifier.

To provide an idempotent interface, you can match this function to {:ok, _}, which will cover all successful flows.

ignored?(phone_number_id)

Returns a boolean value showing whether the phone number is ignored.

Raises Ecto.NoResultsError if no phone number was found with given identifier.

make_primary(passphrase_id, phone_number_id)

Specs

make_primary(
  Legion.Identity.Auth.Concrete.Passphrase.id(),
  Legion.Identity.Telephony.PhoneNumber.id()
) ::
  {:ok, Legion.Identity.Telephony.PhoneNumber.PrioritizationTrait}
  | {:ok, :noop}
  | {:error, :phone_number, :ignored}
  | {:error, :phone_number, :unsafe}
  | {:error, :phone_number, :not_found}
  | {:error, :passphrase, :not_found}
  | {:error, :passphrase, :invalid}
  | {:error, :passphrase, :timed_out}

Makes a phone number primary.

Return values

On successful operation, the function will return {:ok, phone_number}.

Otherwise, it returns

{:ok, :noop}, if phone number was already safe,

{:error, :ignored}, if the phone number was marked as ignored,

{:error, :unsafe}, if the phone number was not marked as safe,

{:error, :not_found}, if there was no phone number with given identifier,

{:error, :no_passphrase}, if there was no passphrase with given identifier.

To provide an idempotent interface, you can match this function to {:ok, _}, which will cover all successful flows.

mark_phone_number_safe(passphrase_id, phone_number_id, opts \\ [])

Specs

mark_phone_number_safe(
  Legion.Identity.Auth.Concrete.Passphrase.id(),
  Legion.Identity.Telephony.PhoneNumber.id(),
  Keyword.t()
) ::
  {:ok, Legion.Identity.Telephony.PhoneNumber.SafetyTrait}
  | {:ok, :noop}
  | {:error, :no_passphrase}
  | {:error, :not_found}
  | {:error, :not_found}
  | {:error, :invalid}
  | {:error, :timed_out}

Marks a phone number safe.

Options

:valid_for: The duration of the new safety trait.

Return values

On successful operation, this will return {:ok, phone_number}.

Otherwise,

{:ok, :noop}, if the phone number was already marked as safe,

{:error, :not_found}, if there was no phone number with given identifier.

To provide an idempotent interface, you can match this function to {:ok, _}, which will cover all successful flows.

mark_phone_number_unsafe(passphrase_id, phone_number_id)

Specs

mark_phone_number_unsafe(
  Legion.Identity.Auth.Concrete.Passphrase.id(),
  Legion.Identity.Telephony.PhoneNumber.id()
) ::
  {:ok, [Invalidation]}
  | {:ok, :noop}
  | {:error, :not_found}
  | {:error, :passphrase_not_found}
  | {:error, :passphrase_invalid}
  | {:error, :passphrase_timed_out}

Marks a phone number unsafe.

Same as mark_phone_number_safe/1, but works opposite.

primary?(phone_number_id)

Returns a boolean value showing whether the phone number is primary.

remove_phone_number(phone_number_id)

Specs

remove_phone_number(Legion.Identity.Telephony.PhoneNumber.id()) ::
  :ok | {:error, :not_found}

Removes a phone number.

safe?(phone_number_id)

Specs

safe?(Legion.Identity.Telephony.PhoneNumber.id()) :: boolean()

Returns a boolean value indicating whether the phone number is marked as safe.

Raises Ecto.NoResultsError if no phone number was found with given identifier.

safe_until(phone_number_id)

Specs

safe_until(Legion.Identity.Telephony.PhoneNumber.id()) :: DateTime.t() | :unsafe

Returns a struct representing safety status of the phone number.

Raises Ecto.NoResultsError if no phone number was found with given identifier.

update_phone_number(phone_number_id, type, number)

Specs

update_phone_number(
  Legion.Identity.Telephony.PhoneNumber.id(),
  Legion.Identity.Telephony.PhoneNumber.phone_type(),
  String.t()
) ::
  {:ok, Legion.Identity.Telephony.PhoneNumber}
  | {:error, :not_found}
  | {:error, :invalid}
  | {:error, :unknown_type}

Updates the type and number of a phone number entry.