Legion.Messaging.Message (legion vdev)

Represents a message requested to be sent to user(s) using a messaging medium.

Types

medium()

Specs

medium() :: :apm | :push | :mailing | :sms | :platform

A medium to send the message by.

Values

:apm: Platform-native push services (i.e. Apple Push Notification Service - APNS, Google Cloud Messaging - GCM, Microsoft Push Notification Service - MPNS, Windows Push Notification Services - WPN). See Legion.Messaging.Medium.APM for more information.

:push: Pseudo-push using local push APIs of devices. Might not be supported on all devices.

:mailing: Sends email using email services.

:sms: Push using Smart Message Service (SMS).

:platform: Sends message using real-time messaging system.

Functions

changeset(struct, params \\ %{})

Validates changeset with given params.

Caveats

:body and :medium fields are required to build a valid changeset.

:subject is required if selected :medium is not :sms.

is_medium(term)
(macro)

Returns true if term is a medium atom (i.e. :apm, :push).

Allowed in guard tests. Inlined by the compiler.

Examples

iex> Legion.Messaging.Message.is_medium(:apm)
true

iex> Legion.Messaging.Message.is_medium(:platform)
true

iex> Legion.Messaging.Message.is_medium(<<1::3>>)
false

iex> Legion.Messaging.Message.is_medium("apm")
false