Legion.Messaging.Switching.Globals (legion vdev)

Provides functions for altering/retrieving global switches to messaging.

This module is NOT transaction-safe.

Enabling/disabling mediums

Suppose you need to disable the a medium globally. You might use enable_medium/2 and disable_medium/2 functions to alter the runtime configuration.

enable_medium(some_user_or_id, :apm)
disable_medium(some_user_or_id, :apm)

Or, rather you can use convenience macros if you require them in your module.

require Legion.Messaging.Switching.Globals

enable_apm_medium(some_user_or_id)
disable_apm_medium(some_user_or_id)

Notice that, the underlying implementation will not insert a new registry entry if value for the setting has not changed. Hence, calling those functions multiple times will not perform any write operation.

Redirecting a medium to another medium

Sometimes you may want to redirect a messaging medium to another medium, probably due to cost reduction and integration maintenance.

You can redirect a medium to another medium with the following call.

redirect_medium(some_user_or_id, :mailing, :apm, for: 3_600)

The above API call will redirect all mailing messages to APM medium. However, while sending a message, you might opt for restricting redirections on such operations.

send_sms_message(some_user_or_id, "this is the message", redirection: :restrict)

The message will not be sent to the user no matter what, what is more, it will throw an error to the user.

Some messages, like one-time-codes, should not be redirected to another medium. The user of the messaging API can also force the actual medium to be run instead of redirection.

send_sms_message(some_user_or_id, "some pretty otc", redirection: :ignore)

If there was a redirection, it will be ignored, although the same rules for enabling/disabling medium for the actual medium will be still applied.

Functions

cancel_redirection_for_medium(user_or_id, medium)

Specs

cancel_redirection_for_medium(
  Legion.Identity.Information.Registration.user_or_id(),
  Legion.Messaging.Message.Medium.t()
) :: :ok | {:error, :no_entry} | {:error, :unavailable}

Cancels redirection setting currently applied for specified medium.

Examples

redirect_medium(user_or_id, :apm, :mailing, for: 3_600) # redirects APM medium to mailing medium for an hour
cancel_redirection_for_medium(user_or_id, :apm) # all redirections for the APM medium are now cancelled

See redirect_medium/4 for making redirections.

disable_apm_medium(user_or_id)
(macro)

Disables in-platform messaging medium.

This macro curries the disable_medium/2 function with corresponding medium.

disable_mailing_medium(user_or_id)
(macro)

Disables mailing medium.

This macro curries the disable_medium/2 function with corresponding medium.

disable_medium(user_or_id, medium)

Specs

disable_medium(
  Legion.Identity.Information.Registration.user_or_id(),
  Legion.Messaging.Message.Medium.t()
) :: :ok | :error

Disables given medium globally.

disable_platform_medium(user_or_id)
(macro)

Disables in-platform messaging medium.

This macro curries the disable_medium/2 function with corresponding medium.

disable_push_medium(user_or_id)
(macro)

Disables push medium.

This macro curries the disable_medium/2 function with corresponding medium.

disable_sms_medium(user_or_id)
(macro)

Disables SMS medium.

This macro curries the disable_medium/2 function with corresponding medium.

enable_apm_medium(user_or_id)
(macro)

Enables APM medium.

This macro curries the enable_medium/2 function with corresponding medium.

enable_mailing_medium(user_or_id)
(macro)

Enables mailing medium.

This macro curries the enable_medium/2 function with corresponding medium.

enable_medium(user_or_id, medium)

Specs

enable_medium(
  Legion.Identity.Information.Registration.user_or_id(),
  Legion.Messaging.Message.Medium.t()
) :: :ok | :error

Enables given medium globally.

enable_platform_medium(user_or_id)
(macro)

Enables platform medium.

This macro curries the enable_medium/2 function with corresponding medium.

enable_push_medium(user_or_id)
(macro)

Enables push medium.

This macro curries the enable_medium/2 function with corresponding medium.

enable_sms_medium(user_or_id)
(macro)

Enables SMS medium.

This macro curries the enable_medium/2 function with corresponding medium.

is_medium_enabled?(medium)

Returns a boolean value indicating if medium is enabled globally.

is_medium_redirected?(medium)

Specs

is_medium_redirected?(Legion.Messaging.Message.Medium.t()) :: boolean()

Returns true if given medium is redirected currently, otherwise false.

redirect_medium(user_or_id, from, to, options \\ [])

Specs

redirect_medium(
  Legion.Identity.Information.Registration.user_or_id(),
  Legion.Messaging.Message.Medium.t(),
  Legion.Messaging.Message.Medium.t(),
  Keyword.t()
) ::
  :ok
  | {:error, :invalid_duration}
  | {:error, :invalid_deferral}
  | {:error, :unavailable}

Redirects a medium to another medium.

Examples

redirect_medium(user_id, :apm, :push) # redirects APM medium to push medium
redirect_medium(user_id, :apm, :platform) # redirects APM medium to platform medium

Timed redirections

You may also redirect a medium to another medium for a given amount of time.

redirect_medium(user_id, :apm, :push, for: 3_600) # redirects APM medium to push medium for an hour

Deferring redirections

Redirections could be also deferred for a given amount of time.

redirect_medium(user_id, :apm, :push, after: 3_600) # redirects APM medium to push medium after an hour

Redirections could be both deferred and timed. The following usage implies both applications.

redirect_medium(user_id, :apm, :push, after: 3_600, for: 6_400) # same redirection, but active after an hour for two hours

Note that redirections can override each other. The user interface for performing redirections should prompt whether the user is aware of overriding an existing redirection.

redirect_medium(user_id, :apm, :push) # redirects APM medium to push medium
redirect_medium(user_id, :apm, :mailing, for: 3_600) # redirects APM medium to mailing medium for an hour, afterwards push redirection will be active until further cancellation

See cancel_redirection_for_medium/3 for cancelling redirections.

redirection_for_medium(medium, timestamp \\ NaiveDateTime.utc_now())

Specs

redirection_for_medium(Legion.Messaging.Message.Medium.t(), NaiveDateTime.t()) ::
  Legion.Messaging.Message.Medium.t() | nil

Returns the redirection medium for the medium in given timestamp, if it is redirected. Otherwise, returns nil.