Legion.RegistryDirectory (legion vdev)

Provides metaprogramming tools to register singleton directories.

Motivation

Global settings directories can be used to gather runtime configuration. One can create a configuration to change the behavior of some modules, or enable/disable it through utility functions.

Suppose you have a function, enable_some_feature/0, to enable a feature at the runtime of the application. To define a backend for this feature in persistence layer, you can leverage registry directories.

Registries

To create a registry directory with a module named SomeSettings, one may use the defregdir/2 provided by this module.

import Legion.RegistryDirectory

defregdir SomeSettings, "messaging_settings"

Upon that call, the macro resolves for defining three modules, namely

SomeSettings.Register,

SomeSettings.RegistryEntry and

SomeSettings.

The first two modules are valid Ecto schemas, you can query them directly.

Repo.all(SomeSettings.Register) # returns all of the register keys

Although it is strictly discouraged, you are also able to query the SomeSettings.RegistryEntry schema.

Retrieving/manipulating settings

The main module, SomeSettings (which was provided to the macro), exports functions for retrieving and manipulating the settings. The functionality expects an authenticated user, or its identifier to perform a manipulation settings (i.e. creating a new entry). The settings do not override each other, but bucketed as an event source, hence you may provide additional features (i.e. deferral, duration), which might likely need more than the last entry created.

Adding registers

Registers can be added at build time, and can be constantly referenced by the other tables. The macro does not perform a migration for the database, indeed, but you can prepare migrations for both key registration and table creation.

For synchronization of the keys, see Legion.RegistryDirectory.Synchronization.

Functions

defregdir(namespace, dirname)
(macro)

Defines a registration directory, resolving to several modules.

{namespace}.Register: An Ecto schema using string keys to refer to the registers.

{namespace}.RegistryEntry: An Ecto schema holds JSON data for a specific key.

{namespace}: Module providing utility functions for manipulating the registry.

The macro also makes use of a base table name, provided by the second parameter, which will resolve to two table names in database, "{namespace}_registers" and "{namespace}_registry_entries", for the two Ecto schemas, respectively.