Legion.Telephony.PhoneNumber (legion vdev)

Declares types and utility functions for working with phone numbers.

Types

host_type()

Specs

host_type() ::
  :fixed_line
  | :mobile
  | :fixed_line_or_mobile
  | :toll_free
  | :premium_rate
  | :shared_cost
  | :voip
  | :personal_number
  | :pager
  | :uan
  | :voicemail
  | :unknown

Describes the type of the phone number.

t()

Specs

t() :: String.t()

Type for the phone number.

Functions

get_number_type(number)

Specs

get_number_type(t()) :: {:ok, host_type()} | {:error, :invalid}

Returns the type of the phone number.

Examples

iex> Legion.Telephony.PhoneNumber.get_number_type("+90 532 111 1111")
{:ok, :mobile}

iex> Legion.Telephony.PhoneNumber.get_number_type("+90 216 111 1111")
{:ok, :fixed_line}

iex> Legion.Telephony.PhoneNumber.get_number_type("test")
{:error, :invalid}

humanize(number)

Same as humanize/2, but uses :international formatting as default.

Examples

iex> Legion.Telephony.PhoneNumber.humanize("+905321111111")
{:ok, "+90 532 111 11 11"}

iex> Legion.Telephony.PhoneNumber.humanize("test")
{:error, :invalid}

humanize(number, format)

Humanizes the phone number in given format. format parameter can be either :international or :national, determining the existence of country code in resulting phone number.

Examples

iex> Legion.Telephony.PhoneNumber.humanize("+905321111111", :international)
{:ok, "+90 532 111 11 11"}

iex> Legion.Telephony.PhoneNumber.humanize("+905321111111", :national)
{:ok, "0532 111 11 11"}

iex> Legion.Telephony.PhoneNumber.humanize("test", :national)
{:error, :invalid}

is_possible_number?(number)

Specs

is_possible_number?(t()) :: boolean()

Returns a boolean value indicating the possibility of validity of the given phone number. Unlike is_valid_number/2, this function validates the phone number by bare lookup of its length.

iex> Legion.Telephony.PhoneNumber.is_possible_number?("+905321111111")
true

iex> Legion.Telephony.PhoneNumber.is_possible_number?("test")
false

is_valid_number?(number)

Specs

is_valid_number?(t()) :: boolean()

Determines validity of the given phone number.

iex> Legion.Telephony.PhoneNumber.is_valid_number?("+905321111111")
true

iex> Legion.Telephony.PhoneNumber.is_valid_number?("test")
false

to_e164(number)

Converts given number to E164-formatted string.

Examples

iex> Legion.Telephony.PhoneNumber.to_e164("+90 532 111 1111")
{:ok, "+905321111111"}

iex> Legion.Telephony.PhoneNumber.to_e164("test")
{:error, :invalid}

to_rfc3966(number)

Specs

to_rfc3966(t()) :: {:ok, String.t()} | {:error, :invalid}

Converts given number to RFC 3966-formatted string.

iex> Legion.Telephony.PhoneNumber.to_rfc3966("+90 532 111 1111")
{:ok, "tel:+90-532-111-11-11"}

iex> Legion.Telephony.PhoneNumber.to_rfc3966("test")
{:error, :invalid}